ViewSets de Activo, Compra y Dividendo con autenticación y filtrado por usuario

+++ portfolio/models.py
@@ -6,7 +6,7 @@ class Activo(models.Model):
     nombre = models.CharField(max_length=50)
     ticker = models.CharField(max_length=5)
     precio = models.DecimalField(max_digits=10, decimal_places=2)
-    cantidad = models.IntegerField()
+    cantidad = models.DecimalField(max_digits=15, decimal_places=8)
 
     def __str__(self):
         return self.nombre
@@ -15,7 +15,7 @@ class Compra(models.Model):
     activo = models.ForeignKey(Activo, on_delete=models.CASCADE) 
     fecha_compra = models.DateField()
     precio = models.DecimalField(max_digits=10, decimal_places=2)
-    cantidad = models.IntegerField()
+    cantidad = models.DecimalField(max_digits=15, decimal_places=8)
 
     def __str__(self):
         return f"{self.activo} - {self.fecha_compra}"
@@ -24,7 +24,7 @@ class Dividendo(models.Model):
     activo = models.ForeignKey(Activo, on_delete=models.CASCADE) 
     fecha_pago = models.DateField()
     div_origen = models.DecimalField(max_digits=20, decimal_places=2)
-    cambio_nominal = models.DecimalField(max_digits=10, decimal_places=2)
+    cambio_nominal = models.DecimalField(max_digits=15, decimal_places=8)
     impuesto = models.DecimalField(max_digits=5, decimal_places=2)
 
     def __str__(self):
